Refactor PlanClient to use pydantic types
