+++ tests/test_sub_clients/test_async_clients/test_plans.py
@@ -3,74 +3,48 @@ from unittest import IsolatedAsyncioTestCase
 import httpx
 from dotenv import load_dotenv
 
-from pypaystack2.sub_clients.plans import AsyncPlanClient
+from pypaystack2.sub_clients.async_clients.plans import AsyncPlanClient
 from pypaystack2.utils import Interval
-from tests.test_sub_clients.mocked_api_testcase import MockedAsyncAPITestCase
-
-
-class MockedAsyncPlanTestCase(MockedAsyncAPITestCase):
-    @classmethod
-    def setUpClass(cls) -> None:
-        super().setUpClass()
-        load_dotenv()
-        cls.wrapper = AsyncPlanClient()
-
-    async def test_can_create(self):
-        response = await self.wrapper.create(
-            name="test plan", amount=600_000, interval=Interval.MONTHLY
-        )
-        self.assertEqual(response.status_code, httpx.codes.OK)
-
-    async def test_can_get_plans(self):
-        response = await self.wrapper.get_plans()
-        self.assertEqual(response.status_code, httpx.codes.OK)
-
-    async def test_can_get_plan(self):
-        response = await self.wrapper.get_plan(id_or_code="710965")
-        self.assertEqual(response.status_code, httpx.codes.OK)
-
-    async def test_can_update(self):
-        response = await self.wrapper.update(
-            id_or_code="PLN_h5yiku242iq0sx5",
-            name="Test plan pro",
-            amount=700_000,
-            interval=Interval.WEEKLY,
-        )
-        self.assertEqual(response.status_code, httpx.codes.OK)
+from pypaystack2.utils.response_models import Plan
 
 
 class AsyncPlanTestCase(IsolatedAsyncioTestCase):
     @classmethod
     def setUpClass(cls) -> None:
         load_dotenv()
-        cls.wrapper = AsyncPlanClient()
+        cls.client = AsyncPlanClient()
 
     async def test_can_create(self):
-        response = await self.wrapper.create(
-            name="test plan", amount=600_000, interval=Interval.MONTHLY
+        response = await self.client.create(
+            name="test plan", amount=600_000, interval=Interval.ANNUALLY
         )
         self.assertEqual(response.status_code, httpx.codes.CREATED)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Plan created")
+        self.assertIsInstance(response.data, Plan)
 
     async def test_can_get_plans(self):
-        response = await self.wrapper.get_plans()
+        response = await self.client.get_plans()
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Plans retrieved")
+        self.assertIsInstance(response.data, list)
+        if len(response.data) > 0:
+            self.assertIsInstance(response.data[0], Plan)
 
     async def test_can_get_plan(self):
-        response = await self.wrapper.get_plan(id_or_code="710965")
+        response = await self.client.get_plan(id_or_code="710965")
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Plan retrieved")
+        self.assertIsInstance(response.data, Plan)
 
     async def test_can_update(self):
-        response = await self.wrapper.update(
+        response = await self.client.update(
             id_or_code="PLN_h5yiku242iq0sx5",
             name="Test plan pro",
             amount=700_000,
-            interval=Interval.WEEKLY,
+            interval=Interval.QUARTERLY,
         )
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
