+++ tests/test_sub_clients/test_sync_clients/test_plans.py
@@ -5,72 +5,46 @@ from dotenv import load_dotenv
 
 from pypaystack2.sub_clients import PlanClient
 from pypaystack2.utils import Interval
-from tests.test_sub_clients.mocked_api_testcase import MockedAPITestCase
-
-
-class MockedPlanTestCase(MockedAPITestCase):
-    @classmethod
-    def setUpClass(cls) -> None:
-        super().setUpClass()
-        load_dotenv()
-        cls.wrapper = PlanClient()
-
-    def test_can_create(self):
-        response = self.wrapper.create(
-            name="test plan", amount=600_000, interval=Interval.MONTHLY
-        )
-        self.assertEqual(response.status_code, httpx.codes.OK)
-
-    def test_can_get_plans(self):
-        response = self.wrapper.get_plans()
-        self.assertEqual(response.status_code, httpx.codes.OK)
-
-    def test_can_get_plan(self):
-        response = self.wrapper.get_plan(id_or_code="710965")
-        self.assertEqual(response.status_code, httpx.codes.OK)
-
-    def test_can_update(self):
-        response = self.wrapper.update(
-            id_or_code="PLN_h5yiku242iq0sx5",
-            name="Test plan pro",
-            amount=700_000,
-            interval=Interval.WEEKLY,
-        )
-        self.assertEqual(response.status_code, httpx.codes.OK)
+from pypaystack2.utils.response_models import Plan
 
 
 class PlanTestCase(TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         load_dotenv()
-        cls.wrapper = PlanClient()
+        cls.client = PlanClient()
 
     def test_can_create(self):
-        response = self.wrapper.create(
-            name="test plan", amount=600_000, interval=Interval.MONTHLY
+        response = self.client.create(
+            name="test plan", amount=600_000, interval=Interval.ANNUALLY
         )
         self.assertEqual(response.status_code, httpx.codes.CREATED)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Plan created")
+        self.assertIsInstance(response.data, Plan)
 
     def test_can_get_plans(self):
-        response = self.wrapper.get_plans()
+        response = self.client.get_plans()
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Plans retrieved")
+        self.assertIsInstance(response.data, list)
+        if len(response.data) > 0:
+            self.assertIsInstance(response.data[0], Plan)
 
     def test_can_get_plan(self):
-        response = self.wrapper.get_plan(id_or_code="710965")
+        response = self.client.get_plan(id_or_code="710965")
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
         self.assertEqual(response.message, "Plan retrieved")
+        self.assertIsInstance(response.data, Plan)
 
     def test_can_update(self):
-        response = self.wrapper.update(
+        response = self.client.update(
             id_or_code="PLN_h5yiku242iq0sx5",
             name="Test plan pro",
             amount=700_000,
-            interval=Interval.WEEKLY,
+            interval=Interval.QUARTERLY,
         )
         self.assertEqual(response.status_code, httpx.codes.OK)
         self.assertTrue(response.status)
